izm_2

+++ src/main/java/hbrnt/dao/UserDao.java
@@ -1,10 +1,13 @@
 package hbrnt.dao;
 
+import hbrnt.models.Auto;
 import hbrnt.models.User;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
+
+import java.util.List;
 
 @Repository
 public class UserDao {
@@ -13,7 +16,7 @@ public class UserDao {
 
     public User findById(int id) {
         Session session = sessionFactory.getCurrentSession();
-        return (User) session.get(User.class, id);
+        return (User) session.load(User.class, id);
     }
 
     public void save(User user) {
@@ -26,9 +29,19 @@ public class UserDao {
         session.update(user);
     }
 
-    public void delete(User user) {
+    public void delete(int id) {
         Session session = sessionFactory.getCurrentSession();
-        session.delete(user);
+        User user = (User) session.load(User.class, id);
+        if(user!=null) {
+            session.delete(user);
+        }
     }
-
+    public Auto findAutoById(int id) {
+        Session session = sessionFactory.getCurrentSession();
+        return (Auto) session.get(Auto.class, id);
+    }
+    public List<User> findAll() {
+        Session session = sessionFactory.getCurrentSession();
+        return (List<User>) session.createQuery("from User").list();
+    }
 }
